feat: improving typos
